Change debugEventsUntilDate in flag model to double (#180)
Unsigned int is not large enough to store millisecond precision timestamps! (Which right now, are about 1.6 trillion ms.)

+++ src/all_flags_state.h
@@ -38,7 +38,7 @@ struct LDFlagState
 
     LDBoolean trackEvents;
 
-    unsigned int debugEventsUntilDate;
+    double debugEventsUntilDate;
 };
 
 /** @brief Returns a pointer to an `LDAllFlagsState` with specified validity.
